requirements.txt
h5py
numpy
pandas
torch
torchtext
torchvision
torchmetrics
wandb
pyspellchecker
pytorch-lightning
pytest
spacy
tensorboard
transformers